Player Animation try
Idea was to animate the scyte draw and to animate putting away the scyte. Ended up being too slow for the pace of the game and I reverted everything. The scyte now just appears and disappears as always.
Task failed successfully lul

Oh and there is a Imapct sound on selling the plants which makes it satisfying (hopefully)